Address simple PR comments.
